pyproject.toml
[tool.poetry]
name = "sentifl-demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
torch = "^2.4.1"
transformers = "^4.44.2"
numpy = ">=1.23.5,<2.0.0"
pandas = "^2.2.3"
keras = "^3.5.0"
tensorflow = "^2.17.0"
konlpy = "^0.6.0"
bs4 = "^0.0.2"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
